+++ route.go
@@ -8,32 +8,34 @@ import (
 )
 
 type HandlerFunc func(request HttpRequest) (response *ResponseMessage)
+type MiddlewareFunc func(request HttpRequest) (response *ResponseMessage, success bool)
 
 type Route struct {
-	route *gin.RouterGroup
+	route      *gin.RouterGroup
+	middleware MiddlewareFunc
 }
 
-func (gp *Route) AddMethod(method string, handler HandlerFunc) {
+func (r *Route) AddMethod(method string, handler HandlerFunc) {
 	method = strings.ToUpper(method)
 	switch method {
 	case http.MethodGet:
-		gp.route.GET("", getHandleFunc(handler))
+		r.route.GET("", r.getHandleFunc(handler))
 	case http.MethodPost:
-		gp.route.POST("", getHandleFunc(handler))
+		r.route.POST("", r.getHandleFunc(handler))
 	case http.MethodPut:
-		gp.route.PUT("", getHandleFunc(handler))
+		r.route.PUT("", r.getHandleFunc(handler))
 	case http.MethodDelete:
-		gp.route.DELETE("", getHandleFunc(handler))
+		r.route.DELETE("", r.getHandleFunc(handler))
 	case http.MethodPatch:
-		gp.route.PATCH("", getHandleFunc(handler))
+		r.route.PATCH("", r.getHandleFunc(handler))
 	case http.MethodHead:
-		gp.route.HEAD("", getHandleFunc(handler))
+		r.route.HEAD("", r.getHandleFunc(handler))
 	case http.MethodOptions:
-		gp.route.OPTIONS("", getHandleFunc(handler))
+		r.route.OPTIONS("", r.getHandleFunc(handler))
 	}
 }
 
-func getHandleFunc(handle HandlerFunc) func(c *gin.Context) {
+func (r *Route) getHandleFunc(handle HandlerFunc) func(c *gin.Context) {
 	return func(c *gin.Context) {
 		body, err := ioutil.ReadAll(c.Request.Body)
 		if err != nil {
@@ -49,17 +51,37 @@ func getHandleFunc(handle HandlerFunc) func(c *gin.Context) {
 		for k, v := range c.Request.Header {
 			headers[k] = v
 		}
+		if r.middleware != nil {
+			message, ok := r.middleware(HttpRequest{
+				Body:    body,
+				Query:   query,
+				Params:  params,
+				Headers: headers,
+			})
+			if !ok {
+				if message != nil {
+					c.JSON(message.statusCode, message)
+					return
+				}
+				c.JSON(http.StatusOK, nil)
+				return
+			}
+		}
 		message := handle(HttpRequest{
 			Body:    body,
 			Query:   query,
 			Params:  params,
 			Headers: headers,
 		})
-		for k, v := range message.headers {
-			for _, h := range v {
-				c.Writer.Header().Add(k, h)
+		if message != nil {
+			for k, v := range message.headers {
+				for _, h := range v {
+					c.Writer.Header().Add(k, h)
+				}
 			}
+			c.JSON(message.statusCode, message)
+			return
 		}
-		c.JSON(message.statusCode, message)
+		c.JSON(http.StatusOK, nil)
 	}
 }