+++ server_interface.go
@@ -3,4 +3,5 @@ package httping
 type IServer interface {
 	NewRoute(group *Route, path string) *Route
 	RunServer() (ServerCloseFunc, chan error)
+	SetMiddleware(middleware MiddlewareFunc) *HttpServer
 }